backend/requirements.txt
# backend/requirements.txt
crewai
crewai-tools
python-dotenv
langchain-community
langchain-cohere
flask
flask-cors